requirements.txt
tkinter==8.6.10
python-dotenv==0.19.2
openai==0.27.0
